feat: implement BATCH_001 tasks via 9-agent Claude-Flow swarm
Implements all 12 tasks from BATCH_001_TASKS.md:

- TASK-001: OpenAI embedding generation (text-embedding-3-small, 768 dims)
- TASK-002: SONA collaborative filtering with SQLx PostgreSQL queries
- TASK-003: SONA content-based filtering with Qdrant vector search
- TASK-004: Ingestion pipeline persistence via ContentRepository trait
- TASK-005: Shared database connection pool (DatabasePool wrapper)
- TASK-006: Auth storage Redis migration with TTL (PKCE, auth codes, device codes)
- TASK-007: PubNub subscribe with MessageHandler trait and long-polling
- TASK-008: Discovery service route wiring (search, trending, content)
- TASK-009: Playback session management with Redis CRUD
- TASK-010: Add num_cpus dependency for worker pool sizing
- TASK-011: Extract shared math utilities (cosine_similarity, dot_product, L2)
- TASK-012: Docker Compose for local dev (PostgreSQL 16, Redis 7, Qdrant)

New files:
- crates/core/src/database.rs - Shared PostgreSQL connection pool
- crates/core/src/math.rs - Vector math utilities
- crates/auth/src/storage.rs - Redis-backed OAuth storage
- crates/ingestion/src/repository.rs - ContentRepository implementation
- crates/playback/src/session.rs - SessionManager with Redis
- docker-compose.yml - Local development infrastructure

+++ crates/auth/src/session.rs
@@ -3,7 +3,7 @@ use redis::{AsyncCommands, Client};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-const SESSION_TTL: usize = 7 * 24 * 3600; // 7 days
+const SESSION_TTL: u64 = 7 * 24 * 3600; // 7 days
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Session {
@@ -77,7 +77,7 @@ impl SessionManager {
             .await
             .map_err(|e| AuthError::Redis(format!("Failed to index session: {}", e)))?;
 
-        conn.expire(&user_sessions_key, SESSION_TTL)
+        conn.expire(&user_sessions_key, SESSION_TTL as i64)
             .await
             .map_err(|e| AuthError::Redis(format!("Failed to set expiry: {}", e)))?;
 
@@ -181,7 +181,7 @@ impl SessionManager {
     }
 
     /// Mark a token as revoked
-    pub async fn revoke_token(&self, jti: &str, ttl: usize) -> Result<()> {
+    pub async fn revoke_token(&self, jti: &str, ttl: u64) -> Result<()> {
         let mut conn = self.get_connection().await?;
 
         let revoked_key = format!("revoked:{}", jti);
